requirements.txt
confluent-kafka==1.1.0
PyYAML==5.1.1
